pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hep-resonance-ad"
version = "0.1.0"
description = "Research platform for CATHODE dimuon anomaly detection"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "HEP Research Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "matplotlib>=3.5.0",
    "pyyaml>=6.0",
    "torch>=1.11.0",
    "awkward>=1.8.0",
    "uproot>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["resonance_ad*"]

[tool.black]
line-length = 100
target-version = ['py38']

